+++ packages/ente-ui/src/App.tsx
@@ -25,6 +25,7 @@ import { Maybe } from "monet";
 import withErrorBoundary from "./hocs/withErrorBoundary";
 import Invitation from "./routes/Invitation/Invitation";
 import InstanceConfigGate from "./components/InstanceConfigGate";
+import { TableStatePersistenceProvider } from "./context/TableStatePersistence";
 
 const { ROTATION_PERIOD } = config.get();
 
@@ -66,9 +67,11 @@ const App: React.FunctionComponent<AppProps> = props => {
               isLoggedIn={authValid}
               purgeStaleData={purgeStaleData}
             >
-              <Drawer>
-                <Routes role={role.orSome(Roles.STUDENT)} />
-              </Drawer>
+              <TableStatePersistenceProvider>
+                <Drawer>
+                  <Routes role={role.orSome(Roles.STUDENT)} />
+                </Drawer>
+              </TableStatePersistenceProvider>
             </AuthenticatedRoute>
           </Switch>
         </>