+++ packages/ente-ui/src/routes/Users/UserTable.tsx
@@ -84,6 +84,7 @@ export const UserTable: React.FunctionComponent<UserTableOwnProps> = props => {
             )
           : undefined
       }
+      persistenceKey="users-table"
     />
   );
 };