+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Don't forget table state when moving off the table ([#312](https://gitlab.com/Skn0tt/EntE/issues/312))
+
 ## [v0.27.1] - 2019-10-31
 
 - Remove serviceworker ([#315](https://gitlab.com/Skn0tt/EntE/issues/315))